server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.18.1
google-generativeai==0.3.2
pydantic==2.6.0
python-dotenv==1.0.1
